+++ src/main/java/com/example/translationchat/domain/report/controller/ReportController.java
@@ -1,10 +1,8 @@
 package com.example.translationchat.domain.report.controller;
 
 import static com.example.translationchat.common.exception.ErrorCode.BAD_REQUEST;
-import static com.example.translationchat.common.exception.ErrorCode.LOCK_FAILED;
 
 import com.example.translationchat.common.exception.CustomException;
-import com.example.translationchat.common.redis.RedisService;
 import com.example.translationchat.common.security.principal.PrincipalDetails;
 import com.example.translationchat.domain.report.service.ReportService;
 import com.example.translationchat.domain.user.entity.User;
@@ -25,7 +23,6 @@ public class ReportController {
 
     private final ReportService reportService;
     private final UserService userService;
-    private final RedisService redisService;
 
     @Transactional
     @PostMapping("/users/{targetUserId}")
@@ -38,19 +35,7 @@ public class ReportController {
         User targetUser = userService.findById(targetUserId);
 
         reportService.validateReport(targetUser, user.getId());
-        String REPORT_LOCK = "REPORT_LOCK";
-        try {
-            boolean reportLocked = redisService.getLock(
-                REPORT_LOCK + targetUserId, 5);
-
-            if (reportLocked) {
-                userService.updateWarningCount(targetUser);
-                reportService.create(targetUser, user.getId());
-            } else {
-                throw new CustomException(LOCK_FAILED);
-            }
-        } finally {
-            redisService.unLock(REPORT_LOCK + targetUserId);
-        }
+        userService.updateWarningCount(targetUser);
+        reportService.create(targetUser, user.getId());
     }
 }
